updated files

+++ mycontacts-backend/controllers/contactController.js
@@ -1,16 +1,17 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel"); 
+
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req,res) =>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler( async(req,res) =>{
     console.log("the req body is :",req.body);
     const { name, email, phone } = req.body;
@@ -22,13 +23,14 @@ const createContact = asyncHandler( async(req,res) =>{
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     res.status(201).json(contact);
 });
 
 //@desc Get new contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler( async (req,res) =>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -40,13 +42,20 @@ const getContact = asyncHandler( async (req,res) =>{
 
 //@desc update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler( async (req,res) =>{
     const contact = await Contact.findById(req.params.id);
         if(!contact){
             res.status(404);
             throw new Error("Contact not found");
         }
+
+        //one user cannot update or access other user contacts data 
+        if(contact.user_id.toString() !== req.user.id){
+            res.status(403);
+            throw new Error("YOU CANNOT UPDATE OTHER USER'S CONTACTS DATA")
+        }
+
         const updatedContact = await Contact.findByIdAndUpdate(
             req.params.id,
             req.body,
@@ -57,13 +66,20 @@ const updateContact = asyncHandler( async (req,res) =>{
 
 //@desc Get all contacts
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler( async (req,res) =>{
     const contact = await Contact.findById(req.params.id);
         if(!contact){
             res.status(404);
             throw new Error("Contact not found");
         }
+
+        //one user cannot delete or access other user contacts data 
+        if(contact.user_id.toString() !== req.user.id){
+            res.status(403);
+            throw new Error("YOU CANNOT UPDATE OTHER USER'S CONTACTS DATA")
+        }
+
         const DeleteContact = await  Contact.findByIdAndDelete(req.params.id);
         res.status(200).json(DeleteContact);
 });
